requirements.txt
streamlit==1.47.1
pydantic==2.11.7
mysql-connector-python==9.3.0
fastapi==0.116.1
pandas==2.3.0
pytest==8.4.1
uvicorn==0.35.0
requests==2.32.3

